pyproject.toml
[project]
name = "hbol"
version = "0.1.0"
description = "Lorentz hyperbolic PPI embeddings (PyTorch + PyG + geoopt)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "torch-geometric>=2.4",
    "geoopt>=0.5.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "tqdm>=4.66",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
umap = ["umap-learn>=0.5"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
